Fix : 컴포넌트 위치 변경,css 수정중
